refactor(CalculateWinner): rename files to match airbnb convension
Issue #14
