+++ client/src/CalculateWinner.js
